TestCaseManagementPlugin:
this change addresses two bugs:
bug 1213 and 1231.

For bug 1213, I've commented out the check to see if the subversion node exists before trying to get the node.  An error is being thrown by the trac api.  Essentially I think they broke something between 10.1 and 10.3.

for bug 1231 i've removed the extra references that were causing an import failure in trac 0.9.6.