Added support for hudson's form-based authentication.
This is based on a modified version of the patch submitted to #6332.
Instead of requiring the user to configure yet another option, Hudson's
403 response is used to trigger the pre-emptive sending of auth info.
This is only very slightly less secure than the config option, and only
so in scenarios where the authentication for Hudson was using Digest auth
but due to some config change Hudson is now returning a 403 - in this case
the plugin will start sending the (essentially cleartext) username and
password which could possibly now be snooped.

This closes #6332 and #6520.