Support of PM values from external relations.  Refs #9691.
While MasterTickets updated the custom fields blocking and blockedby to include dependency information,
it is not in a form that can be used to reliably build a graph of dependant tickets.  This adds support
for getting PM values from external relations such as the `mastertickets` table.

For now this is only used for predecessor/successor but it could be expended to parent/child (e.g.,
using SubTickets plugin's subtickets table) if needed in the future.